bit5 2

+++ app/src/main/java/com/example/part5bitfit/WaterEntry.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "water_entries")
 data class WaterEntry(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val date: String,
-    val glasses: Int
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,  // Auto-generate the primary key
+    val date: String,  // Date when the entry was made
+    val glasses: Int   // Number of glasses consumed
 )
